requirements.txt
Pillow>=9.0.0        # Для работы с изображениями и EXIF
tqdm>=4.64.0         # Прогресс-бар
